improved error handling (read desc)
it looks like due to CORS restrictions, unless you're running firefox, errorevent will contain no information whatsoever, and error will be null.
as such, on the online version of the game, all scripts should be loaded with the crossorigin="anonymous" attribute.
although if I can pack the entire offline version of the game into a single file this hopefully won't be too big an issue.