AUDIO ASSETS ARE WORKING
music by nuke. audio takes up a lot of space in memory so delete() should get it out.
improved the asset format in general too, and also improved crash handling. most of what goes on in draw/update is now wrapped in a try/catch block to hopefully get more details for the crash screen.